examples/genai_rag_solution/components/genai_service/requirements.txt
fastapi
fastapi-restful
fireo
google-cloud-aiplatform
google-cloud-bigquery
google-cloud-storage
langchain_google_vertexai
langchain-core
langchain-experimental
langchain-google-genai
langchain-google-vertexai
langchain-text-splitters
lxml
matplotlib
opencv-python
pillow
poppler
pydantic
pypdf
python-multipart
tiktoken
unstructured[pdf]
uvicorn